Fail safe.

+++ org.jdrupes.vmoperator.runner.qemu/src/org/jdrupes/vmoperator/runner/qemu/QemuMonitor.java
@@ -105,6 +105,9 @@ public class QemuMonitor extends Component {
      */
     @Handler
     public void onStart(Start event) throws IOException {
+        if (socketPath == null) {
+            return;
+        }
         Files.deleteIfExists(socketPath);
         fire(new WatchFile(socketPath));
     }
